Change: Move action method from GmpCommand to EntityCommand class
It's actually only used in EntityCommands. Therefore the method should
be only available for entity command classes. With this change
GmpCommand is empty and obsolete.
